Fix troff warnings in manpages
Newline characters are not allowed in an escape name. Also, troff treats
lines starting with an apostrophe as a control line, that can be
followed by a macro call.

To avoid this, this patch replaces lines containing \n by .br, and uses
\(aq instead of apostrophes on the beginning of a line.

The warnings emitted by troff can be seen using the following command:

LC_ALL=C.UTF-8 MANROFFSEQ='' MANWIDTH=80 \
	man --warnings -E UTF-8 -l -Tutf8 -Z man/voacapl.1 > /dev/null